Add files via upload
This is the first commit to this repository. The included Jupyter notebook includes two examples. The first example is an example of using Bayesian methods to find a point estimate of the weighting of a simulated coin from simulated "experimental" data. The second is an implementation of the Gaussian Naive Bayes Classifier to a toy problem. Specifically, the classification of a point in R^2 as being inside or outside of a collection of circles.  In this example, inclusion within each circle corresponds to a class and not being contained in any circle also defines its own class (i.e., for k circles, there are k+1 classes).